[25073][25092] More robust refreshing of table upon changes
Currently, altering the hierarchy of a work package does not properly
update the table. For example, removing a parent of w1 that isnt visible in
the table causes w1 to refresh in the table (due to cache service), but
the entire table isn't refreshed.

We cannot refresh the entire table for cache service changes, since that
happens quite often in the timeline mode and would cause a peformance
impact.

On the other hand, we don't have the information on _what_ changed in w1
when we subscribe to the cache service.

This PR rather improves on the task of reloading the table after changes
to work packages by making two things explicit:

1. A separate state controlling requests to table refreshing instead of
a rootScope event

2. A service that orchestrates the state to allow moving to/from full
screen to reload the table whenever moving back and requests were made in the full
screen.